pyproject.toml
[tool.poetry]
name = "integracao-hiper-deliverydireto-jobs"
version = "0.1.0"
description = ""
authors = ["Carlos Eduardo <edu-eduardo99@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
Pillow = "^9.2.0"
urllib3 = "^1.26.12"
schedule = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
